requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
chardet==4.0.0
discord.py==1.7.3
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
multidict==6.0.2
PyMySQL==1.0.2
python-dotenv==0.21.0
SQLAlchemy==1.4.41
typing-extensions==4.3.0
yarl==1.8.1
zipp==3.8.1
